requirements.txt
streamlit>=1.22.0
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0
pandas>=1.3.0